Refactor of locations to abs parse and services
Moves services/abstracts back to services/documents
Splits out parse_abs_file
Moves fs_abs to fs_implementation
