+++ browse/config.py
@@ -88,7 +88,7 @@ class Settings(BaseSettings):
     """Enable/disable status service banner."""
 
     DOCUMENT_LISTING_SERVICE: PyObject = 'browse.services.listing.fake'  # type: ignore
-    DOCUMENT_ABSTRACT_SERVICE: PyObject = 'browse.services.abstracts.fs_abs'  # type: ignore
+    DOCUMENT_ABSTRACT_SERVICE: PyObject = 'browse.services.documents.fs_docs'  # type: ignore
     DOCUMENT_LATEST_VERSIONS_PATH: str = "tests/data/abs_files/ftp"
     """Paths to .abs and source files."""
     DOCUMENT_ORIGNAL_VERSIONS_PATH: str = "tests/data/abs_files/orig"
