+++ browse/controllers/abs_page/__init__.py
@@ -26,8 +26,8 @@ from browse.services.search.search_authors import queries_for_authors, \
 from browse.services.util.metatags import meta_tag_metadata
 from browse.services.util.response_headers import abs_expires_header, \
     mime_header_date
-from browse.services.abstracts import get_abs_service
-from browse.services.abstracts.base import AbsException,\
+from browse.services.documents import get_doc_service
+from browse.services.documents.base_documents import AbsException,\
     AbsNotFoundException, AbsVersionNotFoundException, AbsDeletedException
 from browse.domain.identifier import Identifier, IdentifierException,\
     IdentifierIsArchiveException
@@ -36,7 +36,7 @@ from browse.services.database import count_trackback_pings,\
     get_dblp_listing_path, get_dblp_authors
 from browse.services.util.external_refs_cits import include_inspire_link,\
     include_dblp_section, get_computed_dblp_listing_path, get_dblp_bibtex_path
-from browse.services.document.config.external_refs_cits import DBLP_BASE_URL,\
+from browse.services.documents.config.external_refs_cits import DBLP_BASE_URL,\
     DBLP_BIBTEX_PATH, DBLP_AUTHOR_SEARCH_PATH
 
 logger = logging.getLogger(__name__)
@@ -82,7 +82,7 @@ def get_abs_page(arxiv_id: str) -> Response:
         if redirect:
             return redirect
 
-        abs_meta = get_abs_service().get_abs(arxiv_id)
+        abs_meta = get_doc_service().get_abs(arxiv_id)
         not_modified = _check_request_headers(
             abs_meta, response_data, response_headers)
         if not_modified:
@@ -104,7 +104,7 @@ def get_abs_page(arxiv_id: str) -> Response:
         # Dissemination formats for download links
         download_format_pref = request.cookies.get('xxx-ps-defaults')
         add_sciencewise_ping = _check_sciencewise_ping(abs_meta.arxiv_id_v)
-        response_data['formats'] = get_abs_service().get_dissemination_formats(
+        response_data['formats'] = get_doc_service().get_dissemination_formats(
             abs_meta,
             download_format_pref,
             add_sciencewise_ping)
@@ -165,7 +165,7 @@ def _non_critical_abs_data(abs_meta: DocMetadata,
 
     # Ancillary files
     response_data['ancillary_files'] = \
-        get_abs_service().get_ancillary_files(abs_meta)
+        get_doc_service().get_ancillary_files(abs_meta)
 
     # Browse context
     _check_context(arxiv_identifier,
@@ -190,11 +190,11 @@ def _check_request_headers(docmeta: DocMetadata,
 
     last_mod_mime = mime_header_date(last_mod_dt)
     etag = f'"{last_mod_mime}"'
-    
+
     resp_headers['Last-Modified'] = last_mod_mime
     resp_headers['ETag'] = etag
     resp_headers['Expires'] = abs_expires_header()[1]
-    
+
     not_modified = _not_modified(last_mod_dt,
                                  _time_header_parse('If-Modified-Since'),
                                  _get_req_header('if-none-match'),
@@ -300,12 +300,12 @@ def _check_context(arxiv_identifier: Identifier,
     prev_url = None
     if arxiv_identifier.is_old_id or context == 'arxiv':
         # Revert to hybrid approach per ARXIVNG-2080
-        next_id = get_abs_service().get_next_id(arxiv_identifier)
+        next_id = get_doc_service().get_next_id(arxiv_identifier)
         if next_id:
             next_url = url_for('browse.abstract',
                                arxiv_id=next_id.id,
                                context='arxiv' if context == 'arxiv' else None)
-        previous_id = get_abs_service().get_previous_id(arxiv_identifier)
+        previous_id = get_doc_service().get_previous_id(arxiv_identifier)
         if previous_id:
             prev_url = url_for('browse.abstract',
                                arxiv_id=previous_id.id,
