--- browse/services/document/__init__.py
@@ -1,1 +1,0 @@
-"""Browse document service."""
