+++ browse/services/listing/fake_listings.py
@@ -53,11 +53,8 @@ class FakeListingFilesService(ListingService):
             '0704.0276', '0704.0991', '0704.0740', '0704.0473', '0704.0083', '0704.0278', '0704.0006',
             '0704.0735', '0704.0753', '0704.0324', '0704.0600', '0704.0737', '0704.0387', '0704.0659',
             '0704.0432', '0704.0408', '0704.0895', '0704.0088', '0704.0719', '0704.0124', '0704.0508']
-
         items2 = [{'id': id, 'listingType': 'new', 'primary': 'cs.DB'}
                   for id in listings]
-        new_count = len(items2)
-
         items3 = [
             {'id': '0704.0145', 'listingType': 'cross', 'primary': 'cs.DL'},
             {'id': '0704.0075', 'listingType': 'cross', 'primary': 'cs.GT'},
@@ -74,7 +71,6 @@ class FakeListingFilesService(ListingService):
             {'id': '0704.0209', 'listingType': 'cross', 'primary': 'cs.GT'},
             {'id': '0704.0916', 'listingType': 'cross', 'primary': 'cs.DL'},
         ]
-
         items4 = [
             {'id': '0704.0091', 'listingType': 'rep'}, {
                 'id': '0704.0054', 'listingType': 'rep'},
@@ -85,7 +81,6 @@ class FakeListingFilesService(ListingService):
             {'id': '0704.0257', 'listingType': 'rep'}, {'id': '0704.0481', 'listingType': 'rep'}]
 
         lstgs: List[ListingItem] = items2 + items3 + items4  # type: ignore
-
         return {'listings': lstgs[skip:skip + show],
                 'announced': datetime.date(2007, 4, 1),
                 'submitted': (datetime.date(2007, 3, 30), datetime.date(2007, 4, 1)),
