+++ browse/services/util/external_refs_cits.py
@@ -4,7 +4,7 @@ from datetime import date
 from typing import Optional
 
 from browse.domain.identifier import Identifier
-from browse.services.document.config.external_refs_cits import \
+from browse.services.documents.config.external_refs_cits import \
     INSPIRE_REF_CIT_CATEGORIES, DBLP_ARCHIVES, DBLP_START_DATE
 from browse.domain.metadata import DocMetadata
 from browse.domain.category import Category
