+++ tests/test_categories.py
@@ -2,13 +2,14 @@ from hamcrest import *
 import unittest
 
 from tests.test_fs_abs_parser import ABS_FILES
-from browse.services.abstracts.fs_abs import AbsMetaSession
+from browse.services.documents.fs_implementation.fs_abs import FsDocMetadataService
 from browse.domain.category import Category
+
 
 class CategoriesTest(unittest.TestCase):
 
     def setUp(self):
-        self.absService = AbsMetaSession(ABS_FILES+'/ftp', ABS_FILES+'/orig')
+        self.absService = FsDocMetadataService(ABS_FILES+'/ftp', ABS_FILES+'/orig')
 
     def test_categories_for_0906_3421v1_cats(self):
         (id, primary, secondaries) = ('0906.3421v1', 'Combinatorics (math.CO)',
