+++ tests/test_doi.py
@@ -4,12 +4,6 @@ from urllib.parse import urlparse, parse_qs, unquote
 
 from bs4 import BeautifulSoup
 
-from tests.test_fs_abs_parser import ABS_FILES
-from browse.services.abstracts.fs_abs import AbsMetaSession
-from browse.domain.license import ASSUMED_LICENSE_URI
-
-import os
-
 from app import app
