+++ tests/test_inspire.py
@@ -5,7 +5,7 @@ from datetime import date
 from bs4 import BeautifulSoup
 
 from tests.test_fs_abs_parser import ABS_FILES
-from browse.services.abstracts.fs_abs import AbsMetaSession
+from browse.services.documents.fs_implementation.parse_abs import parse_abs_file
 from browse.domain.license import ASSUMED_LICENSE_URI
 from browse.services.util.external_refs_cits import include_inspire_link, \
     get_orig_publish_date, INSPIRE_REF_CIT_CATEGORIES
@@ -23,7 +23,7 @@ class InspireTest(unittest.TestCase):
 
     def test_abs_with_inspire(self):
         f1 = ABS_FILES + '/ftp/arxiv/papers/1108/1108.5926.abs'
-        m = AbsMetaSession.parse_abs_file(filename=f1)
+        m = parse_abs_file(filename=f1)
 
         assert_that( m , is_not( None))
         assert_that( get_orig_publish_date(m.arxiv_identifier), equal_to(date(2011,8,1)))
@@ -39,7 +39,7 @@ class InspireTest(unittest.TestCase):
 
     def test_abs_without_inspire(self):
         f1 = ABS_FILES + '/ftp/math/papers/0202/0202001.abs'
-        m = AbsMetaSession.parse_abs_file(filename=f1)
+        m = parse_abs_file(filename=f1)
 
         assert_that( m , is_not( None))
         assert_that(include_inspire_link( m ), equal_to(False),
